Changed arrow logic, reverted previous commit
Changed arrow logic to show arrows when slideCount < Math.ceil(slidesToShow/2). This accounts for the whitespace to the left of first item in the carousel.

Reverted a737a4c98a4e342f33dcb548f54725c864b917e3 because it screwed up the positioning when centerMode was on and slidesToShow was odd. Adding new slides led to it becoming off-center.